created pokeindex js file and moved app js contents into pokeindex / installed react router
